+++ rambo/app.py
@@ -371,6 +371,7 @@ def up(ctx=None, **params):
         provision (bool): vagrant provisioning flag.
         provision_cmd (str): Command used at beginning of provisioning.
         provision_script (path): Path to script to use for provisioning.
+        provision_with_salt (bool): Flag to indicate provisioning with salt.
         destroy_on_error (bool): vagrant destroy-on-error flag.
         vagrant_cwd (path): Location of `Vagrantfile`. Used if invoked with API only.
         vagrant_dotfile_path (path): Location of `.vagrant` metadata directory. Used if invoked with API only.
@@ -391,8 +392,10 @@ def up(ctx=None, **params):
     params['provider'] = options.provider_option(params.get('provider'))
     params['provision_cmd'] = options.provision_cmd_option(params.get('provision_cmd'))
     params['provision_script'] = options.provision_script_option(params.get('provision_script'))
+    params['provision_with_salt'] = options.provision_with_salt_option(params.get('provision_with_salt'))
     params['ram_size'], params['drive_size'] = options.size_option(
         params.get('ram_size'), params.get('drive_size')) # both ram and drive size
+    params['salt_bootstrap_args'] = options.salt_bootstrap_args_option(params.get('salt_bootstrap_args'))
     params['sync_dir'] = options.sync_dir_option(params.get('sync_dir'))
     params['sync_type'] = options.sync_type_option(params.get('sync_type'))
     params['vm_name'] = options.vm_name_option(params.get('vm_name'))